% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBranchPolygon.R
\name{do_some_cal}
\alias{do_some_cal}
\title{Simple function to test}
\usage{
do_some_cal(x)
}
\arguments{
\item{x}{some number}
}
\value{
some number
}
\description{
Simple function to test
}
\examples{
do_some_cal(5)
}
