% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBranchPolygon.R
\name{getPolygonsfromOSRM}
\alias{getPolygonsfromOSRM}
\title{Title get polygon using OSRM given defalt 0 - 30 increamental 5 mins drive time}
\usage{
getPolygonsfromOSRM(x)
}
\arguments{
\item{x}{dataframe contain some info}
}
\value{
spatial dataframe containing the result
}
\description{
Title get polygon using OSRM given defalt 0 - 30 increamental 5 mins drive time
}
\examples{
\dontrun{
getPolygonsfromOSRM(ds)
}
}
