\name{ppp-class}
\docType{class}
\alias{ppp-class}
\alias{owin-class}
\alias{im-class}
\alias{psp-class}

\title{Virtual class "ppp" }
\description{ Virtual S4 class definition for S3 classes in the spatstat package to allow S4-style coercion to these classes}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\author{ Edzer J. Pebesma }
\keyword{classes}
